Fixing tests and cleaning up mess from fixing calculation util

+++ LifeCountDownTests/DateCalculationUtilTest.h
@@ -31,6 +31,6 @@
 
 @interface DateCalculationUtilTest : XCTestCase
 
--(double) calcCorrectRemainingSeconds:(NSDate*)bDate baseAge:(NSInteger)bAge;
+-(double) calcCorrectRemainingSeconds:(NSDate*)bDate baseAge:(float)bAge;
 
 @end
